move util methods to deployment utils
